requirements.txt
fastmcp==0.1.0